+++ libafl/src/lib.rs
@@ -88,15 +88,16 @@ pub use ctor::ctor;
 #[allow(unused_imports)]
 #[macro_use]
 extern crate libafl_derive;
+pub use libafl_bolts as bolts;
 #[cfg(feature = "libafl_derive")]
 #[doc(hidden)]
 pub use libafl_derive::*;
 
-pub mod bolts;
 pub mod corpus;
 pub mod events;
 pub mod executors;
 pub mod feedbacks;
+pub mod fuzzer;
 pub mod generators;
 pub mod inputs;
 pub mod monitors;
@@ -106,326 +107,8 @@ pub mod schedulers;
 pub mod stages;
 pub mod state;
 
-pub mod fuzzer;
-use alloc::string::{FromUtf8Error, String};
-use core::{
-    array::TryFromSliceError,
-    fmt,
-    num::{ParseIntError, TryFromIntError},
-};
-#[cfg(feature = "std")]
-use std::{env::VarError, io};
-
 pub use fuzzer::*;
-
-#[cfg(feature = "errors_backtrace")]
-/// Error Backtrace type when `errors_backtrace` feature is enabled (== [`backtrace::Backtrace`])
-pub type ErrorBacktrace = backtrace::Backtrace;
-
-#[cfg(not(feature = "errors_backtrace"))]
-#[derive(Debug, Default)]
-/// Empty struct to use when `errors_backtrace` is disabled
-pub struct ErrorBacktrace {}
-#[cfg(not(feature = "errors_backtrace"))]
-impl ErrorBacktrace {
-    /// Nop
-    #[must_use]
-    pub fn new() -> Self {
-        Self {}
-    }
-}
-
-#[cfg(feature = "errors_backtrace")]
-fn display_error_backtrace(f: &mut fmt::Formatter, err: &ErrorBacktrace) -> fmt::Result {
-    write!(f, "\nBacktrace: {err:?}")
-}
-#[cfg(not(feature = "errors_backtrace"))]
-#[allow(clippy::unnecessary_wraps)]
-fn display_error_backtrace(_f: &mut fmt::Formatter, _err: &ErrorBacktrace) -> fmt::Result {
-    fmt::Result::Ok(())
-}
-
-/// Main error struct for `LibAFL`
-#[derive(Debug)]
-pub enum Error {
-    /// Serialization error
-    Serialize(String, ErrorBacktrace),
-    /// Compression error
-    #[cfg(feature = "llmp_compression")]
-    Compression(ErrorBacktrace),
-    /// File related error
-    #[cfg(feature = "std")]
-    File(io::Error, ErrorBacktrace),
-    /// Optional val was supposed to be set, but isn't.
-    EmptyOptional(String, ErrorBacktrace),
-    /// Key not in Map
-    KeyNotFound(String, ErrorBacktrace),
-    /// No elements in the current item
-    Empty(String, ErrorBacktrace),
-    /// End of iteration
-    IteratorEnd(String, ErrorBacktrace),
-    /// This is not supported (yet)
-    NotImplemented(String, ErrorBacktrace),
-    /// You're holding it wrong
-    IllegalState(String, ErrorBacktrace),
-    /// The argument passed to this method or function is not valid
-    IllegalArgument(String, ErrorBacktrace),
-    /// The performed action is not supported on the current platform
-    Unsupported(String, ErrorBacktrace),
-    /// Shutting down, not really an error.
-    ShuttingDown,
-    /// Something else happened
-    Unknown(String, ErrorBacktrace),
-}
-
-impl Error {
-    /// Serialization error
-    #[must_use]
-    pub fn serialize<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::Serialize(arg.into(), ErrorBacktrace::new())
-    }
-    #[cfg(feature = "llmp_compression")]
-    /// Compression error
-    #[must_use]
-    pub fn compression() -> Self {
-        Error::Compression(ErrorBacktrace::new())
-    }
-    #[cfg(feature = "std")]
-    /// File related error
-    #[must_use]
-    pub fn file(arg: io::Error) -> Self {
-        Error::File(arg, ErrorBacktrace::new())
-    }
-    /// Optional val was supposed to be set, but isn't.
-    #[must_use]
-    pub fn empty_optional<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::EmptyOptional(arg.into(), ErrorBacktrace::new())
-    }
-    /// Key not in Map
-    #[must_use]
-    pub fn key_not_found<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::KeyNotFound(arg.into(), ErrorBacktrace::new())
-    }
-    /// No elements in the current item
-    #[must_use]
-    pub fn empty<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::Empty(arg.into(), ErrorBacktrace::new())
-    }
-    /// End of iteration
-    #[must_use]
-    pub fn iterator_end<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::IteratorEnd(arg.into(), ErrorBacktrace::new())
-    }
-    /// This is not supported (yet)
-    #[must_use]
-    pub fn not_implemented<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::NotImplemented(arg.into(), ErrorBacktrace::new())
-    }
-    /// You're holding it wrong
-    #[must_use]
-    pub fn illegal_state<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::IllegalState(arg.into(), ErrorBacktrace::new())
-    }
-    /// The argument passed to this method or function is not valid
-    #[must_use]
-    pub fn illegal_argument<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::IllegalArgument(arg.into(), ErrorBacktrace::new())
-    }
-    /// Shutting down, not really an error.
-    #[must_use]
-    pub fn shutting_down() -> Self {
-        Error::ShuttingDown
-    }
-    /// This operation is not supported on the current architecture or platform
-    #[must_use]
-    pub fn unsupported<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::Unsupported(arg.into(), ErrorBacktrace::new())
-    }
-    /// Something else happened
-    #[must_use]
-    pub fn unknown<S>(arg: S) -> Self
-    where
-        S: Into<String>,
-    {
-        Error::Unknown(arg.into(), ErrorBacktrace::new())
-    }
-}
-
-impl fmt::Display for Error {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Self::Serialize(s, b) => {
-                write!(f, "Error in Serialization: `{0}`", &s)?;
-                display_error_backtrace(f, b)
-            }
-            #[cfg(feature = "llmp_compression")]
-            Self::Compression(b) => {
-                write!(f, "Error in decompression")?;
-                display_error_backtrace(f, b)
-            }
-            #[cfg(feature = "std")]
-            Self::File(err, b) => {
-                write!(f, "File IO failed: {:?}", &err)?;
-                display_error_backtrace(f, b)
-            }
-            Self::EmptyOptional(s, b) => {
-                write!(f, "Optional value `{0}` was not set", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::KeyNotFound(s, b) => {
-                write!(f, "Key `{0}` not in Corpus", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::Empty(s, b) => {
-                write!(f, "No items in {0}", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::IteratorEnd(s, b) => {
-                write!(f, "All elements have been processed in {0} iterator", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::NotImplemented(s, b) => {
-                write!(f, "Not implemented: {0}", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::IllegalState(s, b) => {
-                write!(f, "Illegal state: {0}", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::IllegalArgument(s, b) => {
-                write!(f, "Illegal argument: {0}", &s)?;
-                display_error_backtrace(f, b)
-            }
-            Self::Unsupported(s, b) => {
-                write!(
-                    f,
-                    "The operation is not supported on the current platform: {0}",
-                    &s
-                )?;
-                display_error_backtrace(f, b)
-            }
-            Self::ShuttingDown => write!(f, "Shutting down!"),
-            Self::Unknown(s, b) => {
-                write!(f, "Unknown error: {0}", &s)?;
-                display_error_backtrace(f, b)
-            }
-        }
-    }
-}
-
-/// Stringify the postcard serializer error
-impl From<postcard::Error> for Error {
-    fn from(err: postcard::Error) -> Self {
-        Self::serialize(format!("{err:?}"))
-    }
-}
-
-/// Stringify the json serializer error
-#[cfg(feature = "std")]
-impl From<serde_json::Error> for Error {
-    fn from(err: serde_json::Error) -> Self {
-        Self::serialize(format!("{err:?}"))
-    }
-}
-
-#[cfg(all(unix, feature = "std"))]
-impl From<nix::Error> for Error {
-    fn from(err: nix::Error) -> Self {
-        Self::unknown(format!("Unix error: {err:?}"))
-    }
-}
-
-/// Create an AFL Error from io Error
-#[cfg(feature = "std")]
-impl From<io::Error> for Error {
-    fn from(err: io::Error) -> Self {
-        Self::file(err)
-    }
-}
-
-impl From<FromUtf8Error> for Error {
-    fn from(err: FromUtf8Error) -> Self {
-        Self::unknown(format!("Could not convert byte / utf-8: {err:?}"))
-    }
-}
-
-#[cfg(feature = "std")]
-impl From<VarError> for Error {
-    fn from(err: VarError) -> Self {
-        Self::empty(format!("Could not get env var: {err:?}"))
-    }
-}
-
-impl From<ParseIntError> for Error {
-    fn from(err: ParseIntError) -> Self {
-        Self::unknown(format!("Failed to parse Int: {err:?}"))
-    }
-}
-
-impl From<TryFromIntError> for Error {
-    fn from(err: TryFromIntError) -> Self {
-        Self::illegal_state(format!("Expected conversion failed: {err:?}"))
-    }
-}
-
-impl From<TryFromSliceError> for Error {
-    fn from(err: TryFromSliceError) -> Self {
-        Self::illegal_argument(format!("Could not convert slice: {err:?}"))
-    }
-}
-
-#[cfg(windows)]
-impl From<windows::core::Error> for Error {
-    fn from(err: windows::core::Error) -> Self {
-        Self::unknown(format!("Windows API error: {err:?}"))
-    }
-}
-
-#[cfg(feature = "python")]
-impl From<pyo3::PyErr> for Error {
-    fn from(err: pyo3::PyErr) -> Self {
-        pyo3::Python::with_gil(|py| {
-            if err.matches(
-                py,
-                pyo3::types::PyType::new::<pyo3::exceptions::PyKeyboardInterrupt>(py),
-            ) {
-                Self::shutting_down()
-            } else {
-                Self::illegal_state(format!("Python exception: {err:?}"))
-            }
-        })
-    }
-}
-
-#[cfg(feature = "std")]
-impl std::error::Error for Error {}
+pub use libafl_bolts::Error;
 
 /// The purpose of this module is to alleviate imports of many components by adding a glob import.
 #[cfg(feature = "prelude")]
@@ -564,7 +247,7 @@ pub mod pybind {
     }
 
     crate::impl_serde_pyobjectwrapper!(PythonMetadata, map);
-    crate::impl_serdeany!(PythonMetadata);
+    libafl_bolts::impl_serdeany!(PythonMetadata);
 
     impl PythonMetadata {
         #[must_use]
