+++ libafl/src/mutators/scheduled.rs
@@ -29,7 +29,7 @@ pub struct LogMutationMetadata {
     pub list: Vec<String>,
 }
 
-crate::impl_serdeany!(LogMutationMetadata);
+libafl_bolts::impl_serdeany!(LogMutationMetadata);
 
 impl AsSlice for LogMutationMetadata {
     type Entry = String;
@@ -415,8 +415,9 @@ where
 
 #[cfg(test)]
 mod tests {
+    use libafl_bolts::rands::{Rand, StdRand, XkcdRand};
+
     use crate::{
-        bolts::rands::{Rand, StdRand, XkcdRand},
         corpus::{Corpus, InMemoryCorpus, Testcase},
         feedbacks::ConstFeedback,
         inputs::{BytesInput, HasBytesVec},
