+++ libafl/src/schedulers/powersched.rs
@@ -20,7 +20,7 @@ use crate::{
 /// The n fuzz size
 pub const N_FUZZ_SIZE: usize = 1 << 21;
 
-crate::impl_serdeany!(SchedulerMetadata);
+libafl_bolts::impl_serdeany!(SchedulerMetadata);
 
 /// The metadata used for power schedules
 #[derive(Serialize, Deserialize, Clone, Debug)]
