+++ libafl_bolts/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "libafl"
+name = "libafl_bolts"
 version.workspace = true
 authors = ["Andrea Fioraldi <andreafioraldi@gmail.com>", "Dominik Maier <domenukk@gmail.com>"]
-description = "Slot your own fuzzers together and extend their features using Rust"
+description = "Low-level bolts to create fuzzers and so much more"
 documentation = "https://docs.rs/libafl"
 repository = "https://github.com/AFLplusplus/LibAFL/"
 readme = "../README.md"
@@ -57,8 +57,7 @@ serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 bytecount = "0.6.3"
 
 [dependencies]
-libafl_bolts = { version = "0.10.1", path = "../libafl_bolts" }
-libafl_derive = { version = "0.10.1", path = "../libafl_derive", optional = true }
+libafl_derive = { version = "0.10.1", optional = true, path = "../libafl_derive" }
 
 rustversion = "1.0"
 tuple_list = { version = "0.1.3" }
