+++ libafl_bolts/src/cli.rs
