+++ libafl_bolts/src/serdeany.rs
@@ -65,7 +65,7 @@ where
 #[macro_export]
 macro_rules! create_serde_registry_for_trait {
     ($mod_name:ident, $trait_name:path) => {
-        /// A [`crate::bolts::serdeany`] module.
+        /// A [`crate::serdeany`] module.
         pub mod $mod_name {
 
             use alloc::boxed::Box;
@@ -78,10 +78,8 @@ macro_rules! create_serde_registry_for_trait {
             use postcard;
             use serde::{Deserialize, Serialize};
             use $crate::{
-                bolts::{
-                    anymap::{pack_type_id, unpack_type_id},
-                    serdeany::{DeserializeCallback, DeserializeCallbackSeed},
-                },
+                anymap::{pack_type_id, unpack_type_id},
+                serdeany::{DeserializeCallback, DeserializeCallbackSeed},
                 Error,
             };
 
@@ -581,10 +579,10 @@ macro_rules! create_serde_registry_for_trait {
             {
                 use serde::ser::SerializeSeq;
 
-                let id = $crate::bolts::anymap::unpack_type_id(self.type_id());
+                let id = $crate::anymap::unpack_type_id(self.type_id());
                 let mut seq = se.serialize_seq(Some(2))?;
                 seq.serialize_element(&id)?;
-                seq.serialize_element(&$crate::bolts::serdeany::Wrap(self))?;
+                seq.serialize_element(&$crate::serdeany::Wrap(self))?;
                 seq.end()
             }
         }
@@ -601,7 +599,7 @@ macro_rules! create_serde_registry_for_trait {
     };
 }
 
-create_serde_registry_for_trait!(serdeany_registry, crate::bolts::serdeany::SerdeAny);
+create_serde_registry_for_trait!(serdeany_registry, crate::serdeany::SerdeAny);
 pub use serdeany_registry::*;
 
 /// Register a `SerdeAny` type in the [`RegistryBuilder`]
@@ -612,7 +610,7 @@ macro_rules! register_at_startup {
         const _: () = {
             #[$crate::ctor]
             fn constructor() {
-                $crate::bolts::serdeany::RegistryBuilder::register::<$struct_type>();
+                $crate::serdeany::RegistryBuilder::register::<$struct_type>();
             }
         };
     };
@@ -631,7 +629,7 @@ macro_rules! impl_serdeany {
     ($struct_name:ident < $( $lt:tt $( : $clt:tt $(+ $dlt:tt )* )? ),+ > $(, < $( $opt:tt ),+ >)*) =>
     {
         impl < $( $lt $( : $clt $(+ $dlt )* )? ),+ >
-            $crate::bolts::serdeany::SerdeAny
+            $crate::serdeany::SerdeAny
             for $struct_name < $( $lt ),+ >
         {
             fn as_any(&self) -> &dyn ::core::any::Any {
@@ -656,7 +654,7 @@ macro_rules! impl_serdeany {
     ($struct_name:ident) =>
     {
         impl
-            $crate::bolts::serdeany::SerdeAny
+            $crate::serdeany::SerdeAny
             for $struct_name
         {
             fn as_any(&self) -> &dyn ::core::any::Any {
