+++ utils/libafl_benches/benches/hash_speeds.rs
@@ -3,7 +3,7 @@
 use std::hash::{BuildHasher, Hasher};
 
 use criterion::{black_box, criterion_group, criterion_main, Criterion};
-use libafl::bolts::rands::{Rand, StdRand};
+use libafl_bolts::rands::{Rand, StdRand};
 //use xxhash_rust::const_xxh3;
 use xxhash_rust::xxh3;
